reuse existing messages in the bb status channel
Wiping everything out on reinit is great when everything works 100%.
Sadly, discord hands me 500s ~a dozen or so times per week, and a 500
will cause the bot to go reinitialize everything. This reinit causes
notifications to be fired 100% of the time, so the bot's channel gets
more activity than it strictly needs to have.