remove invisible spaces from emails sometimes
when trying this out, i copy-pasted emails that this bot posted into
an `add-email` command. the bot then proceeded to never work on this
email. after a few minutes of head-scratching, it turns out that my
cleverness with zero-width spaces bit me.

gracefully handle others falling victim to the same issue.