plumbing for greendragon integration; nfc
there are three steps involved in getting greendragon support:
- rambling from an ivory tower in an overly-verbose comment about how to
  support greendragon (check!)
- actually getting data from greendragon (uncheck)
- surfacing that greendragon data through the UI (check!)

this commit makes it so that "all we need to do" now is fetch the data.
that's probably going to involve some code motion and trial and error.
this change is just chasing down compiler errors until there were none.
i <3 decent support for types.