lab: fix race between builds under the same builder
as mentioned in the deleted comment, a single builder can have multiple
workers (aka actual 'machines') working on builds simultaneously. as
written, this code doesn't really know how to cope with this case: it
assumes that if build A is finished before build B, then build A is
using older sources than build B. this is all well and good probably 90%
of the time, but...

so the idea here is to rely on local IDs: if we see a newly-completed
build with a local ID > the local ID of a pending build for the same
bot, we just sorta stash away the finished build for later.