add calendar_check
This is currently a binary that allows for pulling info from the LLVM
calendar. I plan to (partially) library-ize it soon, so it can be
invoked directly from the llvmbb monitor. Should ideally be runnable as
a binary still too, so devs/I can easily troubleshoot when something
goes wrong with it.