WIP: lib-ify calendar_check, and add unfinished calendar_events
This is about as far as I'm going to get today. I'd originally planned
for calendar_check to be a separate binary that spits out JSON and that
the discord bot runs & parses the output of, but it's so much simpler to
just use it as a crate.

Still, that output was nice from a debugging perspective, so allow a
binary to be built.

Things that remain are:
- Actually pinging over discord.
- Probably tests for calendar_events.
- Teaching storage.rs to hold a pinged_calendar_events table so the bot
  is less likely to double-ping.